requirements.txt
paho-mqtt
adafruit_ssd1306
RPi.GPIO
gpiozero